Fix remove not exists image error

+++ src/Infrastructure/Competitions.Persistence/Matches/Handlers/Matches/DeleteMatchConditionsRegulationsHandler.cs
@@ -8,16 +8,17 @@ namespace Competitions.Persistence.Managment.Handlers.Matches
     {
         private readonly IHostingEnvironment _webHostEnv;
 
-        public DeleteMatchConditionsRegulationsHandler ( IHostingEnvironment webHostEnv )
+        public DeleteMatchConditionsRegulationsHandler(IHostingEnvironment webHostEnv)
         {
             _webHostEnv = webHostEnv;
         }
 
 
-        public Task<Unit> Handle ( DeleteMatchConditionsRegulationsEvent request , CancellationToken cancellationToken )
+        public Task<Unit> Handle(DeleteMatchConditionsRegulationsEvent request, CancellationToken cancellationToken)
         {
             string path = _webHostEnv.WebRootPath + request.Path + request.Name;
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
             return Unit.Task;
         }
     }
