+++ src/Infrastructure/Competitions.Persistence/Matches/Handlers/Notifications/DeleteNotificationImageHandler.cs
@@ -8,16 +8,17 @@ namespace Competitions.Persistence.Managment.Handlers.Notifications
     {
         private readonly IHostingEnvironment _webHostEnv;
 
-        public DeleteNotificationImageHandler ( IHostingEnvironment webHostEnv )
+        public DeleteNotificationImageHandler(IHostingEnvironment webHostEnv)
         {
             _webHostEnv = webHostEnv;
         }
 
 
-        public Task<Unit> Handle ( DeleteNotificationImageEvent request , CancellationToken cancellationToken )
+        public Task<Unit> Handle(DeleteNotificationImageEvent request, CancellationToken cancellationToken)
         {
             string path = _webHostEnv.WebRootPath + request.Path + request.Name;
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
             return Unit.Task;
         }
     }
