+++ src/Infrastructure/Competitions.Persistence/Places/Handlers/DeleteActivityPlanFileHandler.cs
@@ -8,16 +8,17 @@ namespace Competitions.Persistence.Places.Handlers
     {
         private readonly IHostingEnvironment _webHostEnv;
 
-        public DeleteActivityPlanFileHandler ( IHostingEnvironment webHostEnv )
+        public DeleteActivityPlanFileHandler(IHostingEnvironment webHostEnv)
         {
             _webHostEnv = webHostEnv;
         }
 
 
-        public Task<Unit> Handle ( DeleteActivityPlanFileEvent request , CancellationToken cancellationToken )
+        public Task<Unit> Handle(DeleteActivityPlanFileEvent request, CancellationToken cancellationToken)
         {
             string path = _webHostEnv.WebRootPath + request.Path + request.Name;
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
             return Unit.Task;
         }
     }
