+++ src/Infrastructure/Competitions.Persistence/Sports/Handlers/DeleteSportImageHandler.cs
@@ -1,11 +1,6 @@
 ﻿using Competitions.Domain.Entities.Sports.Events;
 using MediatR;
 using Microsoft.AspNetCore.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Competitions.Persistence.Sports.Handlers
 {
@@ -13,16 +8,17 @@ namespace Competitions.Persistence.Sports.Handlers
     {
         private readonly IHostingEnvironment _webHostEnv;
 
-        public DeleteSportImageHandler ( IHostingEnvironment webHostEnv )
+        public DeleteSportImageHandler(IHostingEnvironment webHostEnv)
         {
             _webHostEnv = webHostEnv;
         }
 
 
-        public Task<Unit> Handle ( DeleteSportImageEvent request , CancellationToken cancellationToken )
+        public Task<Unit> Handle(DeleteSportImageEvent request, CancellationToken cancellationToken)
         {
             string path = _webHostEnv.WebRootPath + request.Path + request.Name;
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
             return Unit.Task;
         }
     }
